+++ extensions/arango/deployment/src/main/java/io/quarkus/arango/deployment/ArangoDBBuildTimeConfig.java
