+++ extensions/arango/pom.xml
@@ -2,7 +2,6 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
 
     <parent>
         <artifactId>quarkus-build-parent</artifactId>
@@ -10,11 +9,11 @@
         <version>999-SNAPSHOT</version>
         <relativePath>../../build-parent/pom.xml</relativePath>
     </parent>
+    <modelVersion>4.0.0</modelVersion>
 
     <artifactId>quarkus-arango-db-parent</artifactId>
     <name>Quarkus - ArangoDB</name>
     <packaging>pom</packaging>
-
     <modules>
         <module>deployment</module>
         <module>runtime</module>
