+++ extensions/arango/runtime/pom.xml
@@ -3,7 +3,7 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>quarkus-arango-db-parent</artifactId>
+        <artifactId>quarkus-arango-parent</artifactId>
         <groupId>io.quarkus</groupId>
         <version>999-SNAPSHOT</version>
     </parent>
@@ -11,7 +11,7 @@
 
     <artifactId>quarkus-arango</artifactId>
     <name>Quarkus - Arango - Runtime</name>
-    <description>Connect to Arango graph datastore</description>
+    <description>Arango client</description>
 
     <dependencies>
         <dependency>
@@ -35,6 +35,12 @@
         <dependency>
             <groupId>com.arangodb</groupId>
             <artifactId>arangodb-java-driver</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>commons-logging</groupId>
+                    <artifactId>commons-logging</artifactId>
+                </exclusion>
+            </exclusions>
         </dependency>
     </dependencies>
