+++ extensions/arango/runtime/src/main/java/io/quarkus/arango/runtime/health/ArangoHealthCheck.java
@@ -1,17 +1,17 @@
 package io.quarkus.arango.runtime.health;
 
-import com.arangodb.ArangoDB;
-import com.arangodb.ArangoDBException;
-import com.arangodb.entity.ArangoDBVersion;
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
+
 import org.eclipse.microprofile.health.HealthCheck;
 import org.eclipse.microprofile.health.HealthCheckResponse;
 import org.eclipse.microprofile.health.HealthCheckResponseBuilder;
 import org.eclipse.microprofile.health.Readiness;
 import org.jboss.logging.Logger;
 
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
+import com.arangodb.ArangoDB;
+import com.arangodb.ArangoDBException;
+import com.arangodb.entity.ArangoDBVersion;
 
 @Readiness
 @ApplicationScoped
@@ -40,14 +40,13 @@ public class ArangoHealthCheck implements HealthCheck {
      * afterwards.
      *
      * @param arangoDBVersion the result summary returned by the server
-     * @param builder         the health builder to be modified
+     * @param builder the health builder to be modified
      * @return the final {@link HealthCheckResponse health check response}
      */
     private static HealthCheckResponse buildStatusUp(ArangoDBVersion arangoDBVersion, HealthCheckResponseBuilder builder) {
 
         builder.withData("server", arangoDBVersion.getServer());
         builder.withData("version", arangoDBVersion.getVersion());
-
 
         return builder.build();
     }
